# src/Core.Models/WebAPI.Data.Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebAPI.Data",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/Event/{id}": {
      "delete": {
        "tags": [
          "Event"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Event"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          }
        }
      }
    },
    "/v1/Event": {
      "delete": {
        "tags": [
          "Event"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Event"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Event"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Event"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/v1/Person/{id}": {
      "delete": {
        "tags": [
          "Person"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Person"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        }
      }
    },
    "/v1/Person": {
      "delete": {
        "tags": [
          "Person"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Person"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Person"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Person"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Department": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 256,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Event": {
        "required": [
          "chamber",
          "fileId",
          "public",
          "subject",
          "timestampFrom",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "chamber": {
            "type": "integer",
            "format": "int32"
          },
          "timestampFrom": {
            "type": "string",
            "format": "date-time"
          },
          "timestampTo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "public": {
            "type": "boolean"
          },
          "fileId": {
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/EventType"
          },
          "subject": {
            "minLength": 1,
            "type": "string"
          },
          "comments": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          },
          "departmentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "roomId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "room": {
            "$ref": "#/components/schemas/Room"
          },
          "eventChangeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "eventChange": {
            "$ref": "#/components/schemas/EventChange"
          },
          "persons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Person"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EventChange": {
        "required": [
          "chamber",
          "fileId",
          "public",
          "subject",
          "timestampFrom",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "chamber": {
            "type": "integer",
            "format": "int32"
          },
          "timestampFrom": {
            "type": "string",
            "format": "date-time"
          },
          "timestampTo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "public": {
            "type": "boolean"
          },
          "fileId": {
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/EventType"
          },
          "subject": {
            "minLength": 1,
            "type": "string"
          },
          "comments": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          },
          "departmentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "roomId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "room": {
            "$ref": "#/components/schemas/Room"
          },
          "eventId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "event": {
            "$ref": "#/components/schemas/Event"
          },
          "persons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Person"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EventStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "EventType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Person": {
        "required": [
          "description",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/PersonType"
          }
        },
        "additionalProperties": false
      },
      "PersonType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Room": {
        "required": [
          "name",
          "roomNumber"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 256,
            "minLength": 1,
            "type": "string"
          },
          "roomNumber": {
            "maxLength": 32,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}